Fixed bug with damaging one's own spectral weapon.
The test for the agent also being the player caused execution to fall to
the 'else if' block that confirmed that the non-player owner was alive,
then promptly crashed on attempting to print a monster message about the
player.

Now the check that the owner isn't damaging itself is generalized to
monsters as well as players, and is checked outside the is_player check.