Don't consider monsters to have useable ranged weapons while wielding cursed melee weapons
Previously, merely possessing a valid ranged weapon with ammunition was
enough for the game to consider the monster as having a ranged attack,
even if it was impossible for them to actually swap to this weapon.

This probably results in minor improvements in behavior in general
circumstances (such monsters with unuseable ranged weapons not trying
to maintain firing positions when moving around corners), but the most
obvious change is that spriggan assassins who are wielding a cursed
melee weapon, but still in possession of ranged ammo, will continue
to dance away from the player while doing literally nothing at all.