Don't wake up sleeping players by passive poison damage
The net result of it doing so was that poisoned players were nearly
immune to being put to sleep - immediately after falling asleep, they
would take a tick of poison damage and then wake up again. Even if
a stabber was already in melee range, it would not normally have time
to attack even once. There is a decent number of poisoners in Forest
and the stabbers themselves not-infrequently have venom weapons, but
moreover it is bad that deliberately poisoning yourself was a very
effective (if little-known) way to obtain sleep protection.

This is probably something of an Aizul buff, and he is already fairly
scary, but I still think the present behavior is rather bad here and
should be changed.