Moth of Madness
A nasty moth that polymorphs monsters into monsters of the same genus.
For reasons that aren't immediately clear to me, some of the validation
checks are failing so uniques might be generated.

The poly distribution should probably be based on the delta between the
monster's HD and the target HD with a bias toward making it stronger.
Thus an orc is much more likely to become an orc warrior than an orc warlord.

Currently hilarious.