Make monster Mass Confusion's duration against players sane
Because at the moment, Pan attempts to confuse the player for 200
turns in one cast, and is only stopped by the maximum confusion
duration being 40. Not that 40 is a very sane duration for a single
casting of confusion, either (and even an HD 4 caster could achieve
this!)

This commit reduces it to be slightly briefer than plain confusion,
and also decreases the overall power multiplier of the spell, which
was quite high for a monster hex.

As a side note, it sems that actor::confuse and the like have
misleading parameters, since the power argument equates to duration
when applied to players, but does literally nothing when applied
to monsters. There are multiple places that call monster::confuse
with specific formulae for duration, which are all them completely
ignored by the ensuing code. Probably something should be done
about this....