Remove references to racial ammo from throw.cc
This also incidentally removes a free +1 to-hit and to-damage that
almost all ranged monsters got for using "matching" racial launchers
and ammunition (with the race in question being "none"). Also made
monsters only get a bonus to-hit from elven launchers if they are
actually elves.

Racial tags on launchers still look very inconsistent and unintuitive,
and could probably just be removed.