Handle chimera noises
In monster data it changes to S_RANDOM; this is fine for most purposes
for which this is used, e.g. determining whether or not a monster is
silent.

In handle_monster_shouts itself, the shout type is substituted for the
shout of one of the constituent parts.

There is a question mark over do_mon_str_replacements -- in most of the
places it's called from, chimera are irrelevant, and when it's called
from handle_monster_shouts the shout type we already determined is
passed in. So there shouldn't be any strange cases, unless perhaps
someone made something new that used sentient monsters as chimera
components. Would need to refactor mons_shouts to accept a monster
instead of a monster_type to make this any better.