Having taught The Enchantress his best trick, Pan (the unique) departs for parts unknown
Pan's two gimmicks (aside from his confusing name) were Mass Confusion
and Metabolic Englaciation. Unfortunately, only one of these actually
plays extremely poorly with his own band, slowing all of them (with a
greatly liklihood than the player, since you could opt to wear rC!). I
expect that perhaps this wasn't noticed since their Leda's was ALREADY
slowing them. With that gone, he's left with Mass Confusion, which is
now shared by The Enchantress, who I feel is the more important unique
in the zone in any case, and works well with her primary means of action.

At this point, this leaves Pan to be basically a beefier satyr, without
much other differentiation. I am not opposed to the idea of a unique
satyr, but I feel his uniqueness and 'function' should be much stronger
than this if he is to have such a conflated name on top of it. So for
now I am electing to disable him (and the Forest:$ vault that featured
him) until a better name and concept can be found.