Make chimera descriptions more grammary
This slightly improves the the description which now
says things like "The head of a rat, the head of another
rat, and the head of yet another rat." This sounds
slightly less awkward than the previous output. It still
sounds a bit strange sometimes and could probably
be made much more natural with some work.