Rework mon-pick algorithm as a class template
This enables the distributions to be easily used for picking over
enums other than monster_type, and even for arbitrary objects.

The new template is contained in random_pick. It can be used
simply by creating a random_picker<T> and calling its pick
method, or can be subclassed if more complex veto behaviour
is required.