Allow random chimera to be created with MONS syntax
This enables a MONS spec to just be "chimera" and three depth-
appropriate monsters will be used. This can produce bizarre
results but might be useful. It should also allows things
like "place:Zot:5 chimera vs place:Pan chimera" in Arena.

The filters could be refined further to allow random chimera
to not be completely ridiculous.