Change the Apis death effect
Instead of Ely's retribution for killing an apis being minor healing
to nearby hostiles, have her inflict weakness upon you and all nearby
allies.

The healing rarely actually happened (since often if you are killing
an apis, you are not simultaneously injuring other monster) and was
essentially inconsequential when it did. Weakness, on the other hand,
is likely to remain relevant for many characters in a much wider range
of situations while staying true to the spirit of Elyvilon - she is
rebuking you by reducing your own ability to do harm to others.