Macabre Finger Necklace
A warding amulet that provides you with another ring slot.

I rewrote some of the jewellery code, but never re-added the
TOUCH_UI code I removed in _prompt_ring_to_remove. Note that octopodes
never had TOUCH_UI code for their jewellery in the first place. This
should probably be looked at.

Also, the player dumps do not show the amulet
ring on the top, but I do not consider this a problem.