Adjustments to Forest bands
Faun and spriggan defender bands are slightly smaller. Spriggan air
mages now come with a band of air elementals instead of other spriggans.
Spriggan riders come with a band of other riders instead of mixed
spriggans. Druids and berserkers retain their mixed bands, but have a
1 in 3 chance of spawning without a band at all. Mixed spriggan bands
have a larger chance of placing non-plain spriggans, and this list is
weighted to favor air mages (for cloud protection) and disfavor druids
and riders. The Enchantress's band also comes with a gaurenteed air
mage for the same reason. Finally, spirit wolves come in bands of
up to 3, and treants have a 1 in 4 chance of being accompanied by
a druid (who can haste them).