Adjust interaction between snaplasher vines and Fedhas worshippers
It is an open question whether all plant-based attacks should cease
to work against a Fedhas worshipper, but Awaken Trees has done so for
a long time, so continuing to operate under the assumption that plants
under someone else's direct control can still be hostile, this commit
adjusts snaplasher vines to be unneutralizable and also to stop the
player swapping with them, and Fedhas complaining when you hit them.

Possibly the overall behavior of dryads and Fedhas worshippers will
be changed in future, at which point this commit should be revisited.

(As a comedy note, prior to this commit, dryads would cast Awaken Vines
against a Fedhasite, then the vines would immediately start to attack
the dryad, while at the same time Awaken Trees would be attacking and
killing the vines in a perpetual chaotic melee)