Fix layout_cathedral_of_symmetry unintenionally always being polar.
It's supposed to be polar only when placing a central map, but this was
being detected very, very incorrectly.

This leverages functionality first introduced in 5a78294 and
subsequently unused since 075a39c - dgn.map_parameters(), which is
passed as "layout" if the map is being generated as a layout or
"central" if it's specifically being used for a central vault.

I'll have another use for that in another commit or two.