Functionality for allowing overflows to place variable altar counts.
This leverages the map_parameters functionality rediscovered in the
previous commit to pass the tag (or tags) used to select the map when
placing overflow vaults; generic overflow vaults get a
"temple_overflow_generic_#" tag where # corresponds to the number of
vaults the vault is supposed to place. To indicate that the vault
supports a variable number of altars, it can be tagged with all of the
relevant

An ancient vault - minitemple, once upon a time known as minivault_3 -
has been resurrected here to illustrate the concept; it can place up to
fourteen(!) altars, but of course normally places far less than that.

Another more interesting variation on the theme is presented in the new
grunt_overflow_temple_open; it places up to eight altars in patterns
surrounding a central feature.