Limit number of summons per spell type
For each summoning spell there is a limit of how
many summons can be created. Going over this limit
means that the older ones will time out in a turn
or two.

Currently this is applied only to non-necromantic
non-permanent player summons.

This summonsdata[] array can be modified to change
the cap and timeout (i.e. how long old summons last
when they are replaced by new ones) per spell type.

To support a new spell, a called to summoned_monster
is required after the monster has been created, and
the spell needs to be listed in summonsdata[].