Allow spectral weapons to be placed by mons spec, like dancing weapons.
The syntax is 'spectral weapon ; (weapon)'.  Note that the AI is
non-functional when placed by anything but the spell.  No vault should
place a bare spectral weapon and expect it to do anything useful.

This change is mostly so that the monster utility can generate spectral
weapons.  However, the stats are currently only set properly by the
actual casting code.