Correct a major timing issue with monster battlespheres
Since the change to how monster turns are ordered (being interleaved
based on speed/energy instead of consecutively based on index), monster
battlesphere are no longer nearly guaranteed to act after their caster.
This meant that, from the player's perspective, the battlesphere would
fire one turn AFTER the spell that triggered it. Worse, it would be
aimed at where the player was standing that turn, so it was possible
to dodge by moving.

I have addressed this by making monster battlesphere triggering reset
the battlesphere to a full turn's worth of energy, and then reinsert
it into the monster action queue. (This reinsertion is necessary since
it is quite likely it would not be otherwise examined until the following
turn, no matter how much energy it had.) I do not think that cases where
two copies of the same monster end up in the queue at once will have
any effect, as popped monsters with unsufficient or modified action
energy are simply ignored.