Get dlua set_border_fill_type() working again.
This was introduced in 550af32 with the intent that it be used in portal
vaults, and (possibly unintentionally) disabled in c346361.

Three vaults that were intended to use this functionality now do so; the
original vault that used it, trove_treasure_island, never had the call
removed, so it's now working as 550af32 intended it to look.