Give Ogres previously forced to start with a short sword a club instead.
This covers OgHus and OgArs.

No slightly sane ogre is going to take a short blade with them as a
backup (or primary) weapon when going on an adventure; they're going to
take a big stick to hit things with, which is what they do best. OgAses
and OgEns are already intelligent enough to do this - they got clubs
instead of daggers (so do TrAses and TrEns; TrHus and TrArs don't get
weapons, so they're not covered by this change).

Slightly more controversially, I'm making this a +1 club instead of a +0
short sword. I don't think anyone will dispute that clubs are terrible
weapons; it's possibly mechanically equivalent to swap out clubs for
daggers and expect them to perform similarly (not for stabbing
purposes), but a club is going to perform noticeably worse than a short
sword. Really, I think it would make more sense for this to be another
weapon type, but I don't think whips or hammers (the maces classed
between clubs and, well, maces) make thematic sense (and hammers are
used essentially nowhere anyway).