Fix compilation with old (4.1) GCC and libstdc++
We use GCC version <= 4.1 as the criterion for the fix (which is the
same as the similar Solaris fix).  Probably we could use __GLIBCXX__
and/or __GLIBCPP__ to detect this better, but I haven't been able
to test that (in particular, old g++ with a different libstdc++).

This reverts commit 0e1e71dfe8e254a2d925fcb0a7ee6905c4175dd1.  We still
have to do some kind of casting for that particular case, though,
because the first argument is unsigned.