Fix overflow in curse skull monster energy pickup_percent.
Curse skulls got an action energy of 15, meaning their pickup_percent
was 150, outside the range of the int8_t that stores it.  Since none
of the energy usage numbers should be negative, make that a uint8_t
instead.  Also change the other members to uint8_t so we don't need
duplicate overloaded versions of combine().

This wasn't actually relevant to curse skulls, since they had
MONUSE_NOTHING, but it was producing a compiler warning and more
importantly could turn up with some future monster.