Buff Spectral Weapon defenses (AC, EV)
Instead of simply having 2-10AC, 2-10EV linearly over the range of 100
power, it now has 2-12AC, 10-20EV over that range.

In particular, giving it hopefully reasonable EV should make it take
less hits to share with the player.  These numbers are probably still
not enough, but it should be a start.

Also, now it's hit dice are bounded below by 1.  A monster with HD0 has
a number of special cases, such as dying to any hit.  Some of these
might be appropriate for a no-weapon-skill spectral weapon, but I don't
want to figure out all the other side effects it might have.