Don't include MSVC dirent emulation.
We don't actually use it.  Two of contribs, pcre and sdl, reference it from
their autoconfage, so I'm not sure if it's optional or mandatory, and can't
check myself -- not deleting it yet.  In any case, even a cursory look shows
at least two bugs, but I can't think of a case we could call this code,
even not directly.