Recover saves with bad spectral weapons/wellsprings (kilobyte).
MONS_SPECTRAL_WEAPON was inserted before two other enumerators
(0.13-a0-1964-g2fab1c1) and then had a ghost_demon structure added
(0.13-a0-2055-g6cfaa00), both without a tag.  Use the speed to
tell which monster is which; see the comments in unmarshallMonster
for more details.

kilobyte did the first version of the ghost_demon handling, which did
read-ahead to try to guess whether there was actually a ghost_demon
struct in the save.  qoala pointed out that the commit that gave
spectral weapons speed 30 immediately preceded the one that added the
ghost_demon structure (and was pushed at the same time), meaning we can
check use the speed for this, too.