Scale down xp from high tier monsters, further reduce zombie xp
This is part of a multi-pronged approach to curbing xp through midgame,
particularly the later parts of it. The amount of xp available in a normal
3-rune game has increased substancially over the last several versions (by
30% or more) and much of this is concentrated in late D. A major factor in
this is the greater frequency of major vaults, but these vaults are a source
of considerable tactical interest and challenge, so it seems bad to address
the growth in average player power by simply pruning them. Moreover, it also
seems a bad thing to dilute interesting threats in branches with chaff simply
to avoid giving too much xp (somthing which is currently done in several
places)

Uniformly increasing the (already existing) diminishing returns on xp from
larger threats reduces the total xp in the game by a noticable amount (though
it is still higher than it used to be) and moreso affects late D and V more
than Lair rune branches, and earlygame very little at all.

Midgame is generally rendered less interesting by an excess of player power,
so the aim of these commits is to curtail it (reasonably so) without affecting
earlygame (which is already plenty hard), and also to make doing Lair rune
branches earlier more desireable, as they become proportionally more rewarding
at the same time that reduced player power makes late D more dangerous.