Some adjustments to wad_woods monster sets
Most importantly, don't place porcupines as early as D:2. Since the
spine change, it is unreasonable to expect many characters of that
level to kill one. Now the vaults instead place green rats before
absdepth 8. Also remove the chance of low tier tengu when it shows
up in Forest, and give a smaller chance for some other monsters to
appear there.

This is probably more a bandaid than anything else, as gnolls in
Lair is still rather odd (and irrelevant) so I think the spawn lists
could use a more thorough look, but at least this should stop certain
unkind earlygame deaths.