Fixes help menu scrolling when at the first line
The menu code was jumping to the previous entry instead of the actual
hotkey'd entry. This could be a MEL_TITLE which leads to some broken
behavior in the help menu.

The most noticeable is in webtiles, where you couldn't use the down arrow
to scroll when at the top of a help menu.

In console a blank line showed up at the top of the help page, but you
couldn't scroll back to it.

This patch fixes it by jumping to the correct entry (the entry after the
MEL_TITLE).
(cherry picked from commit 280b7835a7bc0c21b61ce711077aaaf040d7bdc9)