Revamp player xp draining
Instead of permanently reducing player xp (and xl), draining attacks
now inflict a temporary reduction in all skill levels (similar to Ash
wrath) which is worked off by gaining xp. The xp used to overcome the
penalty is not actually deducted, so draining does not reduce long-term
xp totals. Draining causes both a flat and (smaller) percentile reduction
(meaning that higher level skills are somewhat more strongly affected
than low ones).

The amount of draining inflicted by an attack scales with the damage
taken, meaning that low EV, high AC characters will be less
proportionately vulnerable to draining than their inverse counterparts,
at least compared to the present situation.

The xp cost of removing a point of draining decreases the more heavily
drained the player is, so that it is a bit easier to dig yourself out
of a deep hole if you manage to end up in one.

The numbers are provisional, but should hopefully be fairly sane. Ideally,
I think it would be good if draining had a noticable enough impact to
encourage wearing rN more often than is presently done, as a counter to
it, but let's see how it plays.

For the time-being, draining still works the same against monsters as
it always has.