Rename Rings of Life Protection to Rings of Positive Energy
Life Protection was a somewhat misleading name (especially for beginners)
since it implies it might help save you from dying when in fact it does
very little on that front in general. Moreover, it was named very
differently from the armour ego which has the exact same effect. And
now it is not.