Increase damage of monster Bolt of Draining
This did appreciably less damage than other high level monster bolts
(by a margin greater even than the player spell), to the point that
mystic blast was equivalent in damage (yet much more accurate!). With
the buff to rN's effect against it, I have increased the base damage
of the spell a little in compensation.

Also, reduce the damage-scaled component of the resulting player xp
drain so that this does not boost that aspect of the spell.