Don't crash when a dead monster hurts a spectral weapon.
See for example !lm simm crash 8 -log .  The deferred_damage_fineff that
this created no longer had access to the dead monster (it was cleaned up
already, so not in the mid cache), so passed NULL as the agent to hurt().
That was handled "properly" (without crashing) by monster::hurt, but not
by player::hurt.

For now, avoid the crash at the expense of a less-than-descriptive death
message ("Slain by posthumous revenge").  For now the only
deferred_damage_fineff that can affect a player (spectral weapon damage
sharing) is non-fatal, so the message shouldn't actually happen.
Handling this situation properly would probably require postponing
cleanup of dead monsters until after fineffs run, and that sounds quite
likely to cause many new bugs.