Remove player clinging
As a player mechanic, it's almost never relevant and in the few cases
where it can be relevant it's quickly obsoleted by fairly common
sources of flight, as well as being unnecessarily complicated (no
diagonal clinging for some reason?). As a race-specific mechanic it
also results in a lot of annoying message spam ("You open the door. You
fall off the door.").