Revert monster gargoyle clinging/holiness changes
The holiness changes made no sense - player gargoyles are nonliving, so it
made the monsters more "alive" than the players. Removing flight and replacing
it with clinging did not factor in that some vaults place gargoyles above lava,
and assume that they are able to fly. Making such changes to the monster just
on the basis of symmetry when the player species does not yet have a final
concept seems like a bad idea, and we should probably keep the two separate
until that point.

This reverts commits 959fb166256fd6ea7, 4050f98598dd9164, 02a3e516c11f06997ac.