Adjust Statue Form for gargoyles
Mostly this means that they get to keep (some) of their natural AC
in statue form, like draconians do in dragon form (it doesn't make
a lot of sense for stone to turn into harder stone and become less
resilient in the process).

Also adjust a message for the form ending (you're not very fleshy).