Make deep elf priests more distict
In the same vein as the last commit, this gives deep elf priests a
unique spell to further differentiate themselves from 'orc priests
who sometimes have a bow'.

Malign Offering is a new spell that offers up the life force of the
target to heal nearby allies of the caster. It is resisted with rN
and requires line of fire, but ignores AC in the manner of drain life.
Allies of the caster within LoS of the target are healed proportional
to the damage inflicted.

While deep elf high priests don't necessarily need any similar boosts,
I have given them the spell as well, for consistency's sake. Both
priests lose their minor healing spell in the process (and normal
priests also lose cantrip)

(Also give plain priests a very slight HD boost)