Remove Deep Elf Soldiers (plus band/vault adjustments)
Stat-wise, these were virtually identical to orc wizards in every
way (except for somewhat better equipment), very weak for the branch
they appeared in, and simply a plain worse version of another common
(and not-so-scary) monster already in that same branch.

Vault-wise, most uses of soldiers have been replaced with fighters.
Sometimes the total number of them was decreased a little, and in one
case a very early vault was shifted deeper, though most cases seem
like they could handle an outright buff without much of consequence
changing.

The contents of the deep elf bands that used them has been altered
somewhat. Fighter and knight bands are somewhat smaller, with fighters
taking up the largest part of the slack for the missing soldiers, but
there is also an increased presence of (improved) mages, along with a
few other tweaks. On the whole, there should be somewhat fewer elves
together, but the ones that are present should be more notable.