Don't crash on vampire blood acquirement.
We were checking whether the object was useless while it still had
base_type OBJ_FOOD but sub_type POT_BLOOD; the latter subtype (27)
is out of bounds for food, causing a crash. See for example:
!lm okayed vpas crash 1 -log (this was Xom "acquirement").

Fix this by overriding the object's base class in a different location.

This will need to be revisited if vampires are ever able to worship
a god who hates blood: currently blood acquirement would give random
potions in that situation.