Don't crash on Malign Offering a monster to death.
... and don't make the healing depend on blindness and invisibility.

We were using an actor_iterator around the victim, but after the victim
took (perhaps lethal) damage.  If the caster was invisible, this meant
checking whether the dead victim can see invisible and thereby
triggering an assertion.  See for example:

  http://dobrazupa.org/morgue/dialectric/crash-dialectric-20130715-183112.txt
    !lm dialectric sprint crash 1 -log

Instead, pass in the victim's los_no_trans.  This is the usual line-of-
effect check, blocked by glass and smoke but still counting invisible
monsters.  This should not only fix the crash, but also make MO's
healing a little more consistent with other effects.

If it really is desired to allow soul juice to pass through glass,
change victim->get_los_no_trans() to victim->get_los().