New player spell: Searing Ray
This is a level 2 pure conjurations spell, intended to replace Force
Lance in the conjurer starting book.

While Force Lance had a decent concept, its role in the book was sort
of awkward. Its effect was unreliable, usable only at extreme close
range, and cost twice as much as magic dart at a time where mp is at
a premium, yet didn't actually hurt twice as much (and could miss).
For many characters, by the time you would have enough mp and spellpower
to make good use of it, you had already outgrown it for better spells.
This spell is aimed to better fill a transitional niche between magic
dart and the stronger conjurations, without rendering the former
obsolete, either.

Searing Ray is a close range spell (though not quite AS close range)
which fires a continuous ray of energy that starts out weak, but grows
in power over time as it is channelled. Mechanically, the caster gains
a 'Ray' status once they fire it, and each subsequent turn that they
perform no other action, the searing ray refires along the original beam
path. It will do this 3 times, each more powerful than the last, and at
the most powerful stage, the ray becomes a penetrating beam.

1 mp is charged for each turn the ray is maintained (and it will end
prematurely if you run out), and the player is offered a warning prompt
and chance to cancel if the beam path becomes unsafe (ie: an ally wandered
into it) but it cannot be reaimed without starting the channeling from
the beginning.

The final stage is quite powerful for a level 2 spell, but it also
requires remaining stationary at a close range to your target for 4 turns,
and for them not to move out of the beam path during this (and also costs
5 mp in total). The initial stages, in contrast, are weaker than other
level 2 conjurations like throw frost or flame.

Numbers probably need more adjusting, but I think the basic concept fills
the level 2 slot in the Book of Conjurations better than Force Lance did.