Raise Force Lance to level 5, buff accordingly
Let's see if the spell can see a little more general utility use at
a higher level. Range is increased from 3 to 4, and damage is boosted
considerably (though it is still noticably lower than similar level
conjurations).

It's possible it will still be underwhelming and/or impractical
compared to other options, but we'll consider it provisional, pending
actual playtesting.