Spellbook adjustments
Swap Force Lance for Searing Ray in the Book of Conjurations, add
upgraded Force Lance to the Book of Wizardry, replacing Lightning Bolt.

Also add Fulminant Prism to the Book of Power, replacing Poisonous
Cloud (which already has two other sources). This is a fun spell that
still has midgame usefulness, so it may see a bit more play if it's
more common.