A .de quote for the Temple, erase a copy of .pl misplaced as .en.
Thanks for MagicTigger for finding a good .de translation; I've been
unable to find an .en one, and I'm unable to translate poetry myself,
especially if one wants an archaic feel.  (Yeah, yeah, when Krasicki
wrote it, it was not archaic.)

Not sure why it landed in .en, there might be a problem with Transifex
sync process when a quote exists in a language but not in .en.