Don't place boring traps randomly.
"Boring" is defined here as without a strategic meaning.  Teleport is
good as it's likely to put you in trouble, shafts move you to a whole
new level.  Not sure about alarm traps: out of vaults, they might still
be mostly meaningless.

For now, all traps still exist, for placing in vaults and for ZotDef.
Removal in the future would be good...