Adjust the list of spells which trigger battlesphere
Several times it has been suggested that it work with directed damaging
spells in general and not just directed CONJURATIONS (which had previously
excluded such conjuration-like things as sandblast and non-damaging things
like mephitic cloud).

This commit makes meph no longer trigger it, but adds sandblast, freeze,
airstrike, and sticky flame to the list.

Also I have additionally removed orb of destruction, since I feel the
delayed nature of the large orb is sort of awkward in consort with the
immediate battlesphere blast, but I am quite open to reverting this if
someone disagrees.