Give a better message when the terminal is too small for the layout.
It was possible to get e.g. "Terminal too small (80,25); need at least
(79,24)".  Now we have a slightly different message when the problem
is the layout rather than the absolute (79,24) limit; and we include
the actual required size.

The required size reported by this commit isn't perfect because there
doesn't seem to be a good way of asking for the true minimum size along
a sufficiently-large axis (leftover_x() is zero even though things could
be shrunk).  We currently report the absolute minimum for such axes,
even though the true limit could in fact be somewhere between that and
the current size.