Let rltiles work with SDL2.
(Makefile changes not included for now.)

For android, we currently use a modified copy of a random alpha version of
SDL trunk, which needs a separate code path.  I have no such toolchain to
test; if it identifies itself as 2.0 rather than 1.3 then we have a problem.

Note that this is merely the "rltiles" tool rather than Crawl itself;
the latter seems to require quite significant porting.  The two next
commits are but a start of that task.