Do not generate hammers unless defined in a vault
This sets their item rarity to 0 and disallows monsters to spawn with a
hammer as weapon. Yiuf was not harmed in the process.

Their weight of 20 has been redistributed to 19 as follows:
- hand axes (+6) which have the exact same stats
- clubs (+5)
- whips (+5)
- flails (+3)

I also considered a mapping like clubs (+10), maces (+5), flails (+3),
whips (+2) instead but ultimately went for hand axes as suitable filler
because of the exact same stats.