Speed 10 for jellies (was: 9)
With speed 9, very tedious tactics like kiting the jelly around and tossing
a bunch of stones at it over and over again was pretty much "optimal", but
took a long time due to their HP and the small difference in movement speed.
The summary is that they now are speed 10 (same as normal speed players,
and most other monsters) but other attributes have been changed to account
for this.

A brief overview of those compensations:
- reduce their base damage to 3 (was: 8)
- lower their HP to 12-21 (was: 15-30)
- increase their experience value to 42 (was: 22)
- make them slightly less common in D (rarity 514, was: 690).

Another change I considered was reducing their acid damage from the quite
high 7d3 to 7d2. That change however required yet another special case in
handling AF_ACID attacks which I could not quite convince myself to do.
If somebody else either improves AF_ACID or hacks in the special case, it
might make sense to increase jelly base damage a bit again.

Expanding on the "why" a bit more:
- generally agreed upon that kiting should not be encouraged
- speed 9 is the worst offender in that it takes forever to open up gaps
- fleeing, i.e. not engaging, monsters in the current situation also is a
  form of kiting that is more tactically interesting with speed 10