Buff Giant Fireflies
Giant fireflies now gain the ability to flash a signal beacon. This is
somewhat similar to generating noise, though with a few significant
differences. Firstly, it alerts only creatures which are already awake,
with the exception of other fireflies and spriggan riders, who are keyed
to be alerted by these signals. Secondly, the signal gives more precise
locational information than noise does, making it easier for monsters to
hone in on the specific position it occurs.

Also, double their pathetic melee damage and increase hp somewhat.

The hope is that this will make the fact that spriggan riders sometimes
leave a firefly upon death ocassionally meaningful (as opposed to
completely meaningless at present)