Fix spectral wolf pacification bug
Spectral wolves continue attacking after being pacified because in
mon-behv.cc, behavior due to haunting overrides behavior due to
paficication. I added a line to mons_pacify to remove the haunting
enchantment. (Haunting causes a monster to move toward and attack
the source of the enchantment, so it should never be relevant once
a monster is pacified.)