Use gold instead of bfd if possible.
In Debian, gold used to be the default linker when both were installed; after
a recent upload of binutils this is no more.  That's a big regression, as
bfd is slower by a factor of 3-4 in most cases, and over 80 times when bfd's
massive memory use makes it hit the swap.

On the other hand, gold is now installed in the base binutils package, and
can be switched to with -fuse-ld=gold.  That's said to be the "upstream
default".  This syntax is not recognized by older gcc nor clang and thus
has to be detected on runtime.