Prevent orb apportation in Sprint
Since it turns out to be a good idea to apport the orb all the way
across the map after clearing out earlier areas, if possible. This
isn't so much the case in the regular game where Zot:5 enemies might
still show up but maybe it'd be simpler/more consistent to just make
the orb always apportation immune.