Remove ISFLAG_BEEN_IN_INV.
It was once used for tracking lost runes (4ba5811), but that's been
irrelevant for a while now.  The vestigial code:

 * Only checked the flag for unrandom artefacts, or to determine
   whether to destroy an item with wizmode &G.
 * Only set the flag for items the player started with (or, prior to
   this commit's parent, blood that coagulated in the inventory).

Since players don't start with unrands, it was relevant exactly never
in normal games.  It was only rarely relevant for &G, since monsters
aren't likely to have player starting inventory that wasn't also
ISFLAG_DROPPED or ISFLAG_THROWN (both of which are still checked).