Don't render gates as doors after sealing/unsealing.
We were sealing one door cell at a time; tile_init_flavour considered
the just-sealed door to be separate from the not-yet-sealed ones.
Saving and restoring re-initialised the tile flavour, correcting the
rendering bug (until the gate unsealed and the same thing happened).

Correct this by considering a sealed door to be the "same door" as an
adjacent closed or runed door (and vice versa).  Another option,
suggested by DracoOmega, would be to make a separate tile_init_flavour
pass after we finish sealing/unsealing.