Support for multi-level sprints.
The depth of the sprint is set by calling crawl.sprint_depth() from a
hook in the main Sprint map. The function takes two arguments - a branch
and a depth (in most cases branch will be 'D', but there are some other
applications of this that may come into play if/when sprints come up
that involve other branches, such as a Zigsprint that uses actual Zig
code).

Vaults for later Sprint levels are tagged "sprint_level" and with the
map ID of the sprint that contains it, like tutorial maps but with the
extra "sprint_level" tag.