Make Zigsprint use actual ziggurat code.
There are a couple of sprint-specific checks this introduces - the Orb
gets placed between the exits on Zig:27, for example - which leverage a
new crawl.is_sprint() Lua function.

I've taken great care to preserve the monster sets and loot generation.

Note that the player can leave the Zig at any time in this version
(removing the exits causes portal vault exit checks to fail which I'm
not willing to work around at this time), though if they do so without
the Orb they're obviously no longer capable of winning.