Small adjustments to a couple of encompass vaults.
Namely, grunt_twisty_little_passages and grunt_runaround.

From what I've heard, the vaults play well when players are clearing
them, but it becomes a tremendous hassle to navigate back through them
after they've been cleared. (This was part of the original motivation
for twisty_little_passages; no, I don't know why I thought it was a good
idea at the time.)

twisty_little_passages now contains two rock chunks across from the
respective pairs of up stairs, and one wand of digging on the interior
side of one of the two pairs of rocks - the idea being that once you've
cleared to that point in the vault, you can just dig through the rock
for a much faster exit. (Of course, players prepared enough to be able
to dig when entering the vault can have a much easier time if they just
want to get to a downstair as soon as possible).

runaround now has each tier generate two doors to the next tier, except
leading into the central chamber, which still only has one door. This
should help make the map more tactically interesting and reduce the
average distance a player needs to travel from stair to stair.